refactor: runtime event listeners in `background.ts`
